Refactor the code that loads a grade book from a file so that it can also be used for saving a grade book to a file.
